+++ Venmo/src/main/java/com/braintreepayments/api/venmo/VenmoClient.kt
@@ -97,7 +97,11 @@ class VenmoClient internal constructor(
         context: Context,
         authorization: String,
         returnUrlScheme: String? = null
-    ) : this(BraintreeClient(context, authorization, returnUrlScheme))
+    ) : this(BraintreeClient(
+        context = context,
+        authorization = authorization,
+        deepLinkFallbackUrlScheme = returnUrlScheme
+    ))
 
     /**
      * Start the Pay With Venmo flow. This will return a [VenmoPaymentAuthRequestParams] that